refactor: Move `file_database.*` to `core`
