+++ libs/deeplog/src/dplx/dlog.test.cpp
@@ -13,7 +13,7 @@
 #include <fmt/format.h>
 
 #include <dplx/dlog/bus/mpsc_bus.hpp>
-#include <dplx/dlog/file_database.hpp>
+#include <dplx/dlog/core/file_database.hpp>
 #include <dplx/dlog/llfio.hpp>
 #include <dplx/dlog/log_fabric.hpp>
 #include <dplx/dlog/sink.hpp>
