+++ libs/deeplog/src/dplx/dlog/core/file_database.cpp
@@ -5,7 +5,7 @@
 //         (See accompanying file LICENSE or copy at
 //           https://www.boost.org/LICENSE_1_0.txt)
 
-#include "dplx/dlog/file_database.hpp"
+#include "dplx/dlog/core/file_database.hpp"
 
 #include <bit>
 #include <chrono>
