+++ libs/deeplog/src/dplx/dlog/core/file_database.test.cpp
@@ -5,7 +5,7 @@
 //         (See accompanying file LICENSE or copy at
 //           https://www.boost.org/LICENSE_1_0.txt)
 
-#include "dplx/dlog/file_database.hpp"
+#include "dplx/dlog/core/file_database.hpp"
 
 #include <catch2/catch_test_macros.hpp>
